Type Definition chess_ai::engine::figures::MovementSet

pub type MovementSet = [[(i8, i8); 7]; 8];

Contains all moves a figure can make. (119, 119) and [(119, 119); 7] are filler moves to fill the array to the right size. They should be ignored.

A movement set contains multiple movement sub sets. Movement sub sets are independent of each other but if a figure cannot make one move of a movement sub set, it also cannot do any of the following moves of that movement sub set.

Example

[
    [
    (1, 1),
    (2, 2),
    (3, 3),
    ...
    ],
    [
    (-1, -1),
    (-2, -2),
    (-3, -3),
    ...
    ]
]

Let’s say the figure can move (1, 1), but not (2, 2). Then only (1, 1) is a possible move of that sub set. (3, 3) is not possible, because could only be possible if (2, 2) is possible.

The next sub set (-1, -1), (-2, -2), (-3, -3) could still be possible.