Module chess_ai::io

Functions

get_str

Konvertiert die gegebene figures::Figure zu einem colored::ColoredString.

handle_move_request

Funktion, die versucht einen vom Spieler gewünschten Zug auszuführen.

handle_possible_moves_request

Funktion, die alle möglichen Züge für eine Figur findet und in die Konsole ausgibt.

is_in_movement_set

Überprüft, ob eine Koordinate in einem MovementSet enthalten ist

parse_input

Parst den vom Spieler gegebenen Input mit der Gewünschten Aktion, falls möglich.

print_board

Gibt ein Spielbrett aus.

print_board_with_movements

Gibt ein Spielbrett aus und markiert die gegeben Zugmöglichkeiten und die ziehende Figur.

turn

Zug-Funktion für Menschen.

validate_coordinate

Validiert eine Koordinate.